requirements.txt
Pillow>=10.0
PyMuPDF>=1.24
pydantic>=2.0
python-dotenv>=1.0.1
openai>=1.30
PyQt6>=6.4
pyqtgraph>=0.13.0
soundfile>=0.12
pytest-qt>=4.2.0
pytest-mock>=3.12.0

ruff>=0.5
black>=24.0
mypy>=1.8
pytest>=8.0
coverage>=7.0
